speech-synthesis-backend/requirements.txt
fastapi
uvicorn[standard]
httpx 
python-dotenv
crewai
langchain_mistralai
psycopg2-binary
sqlalchemy
alembic
PyJWT==2.8.0